moved SearchBox.test.js into SearchBox dir
